Refactor utils file and place all authorization code in auth file
